% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_request.R
\name{CatalogRequestBuilder}
\alias{CatalogRequestBuilder}
\title{CatalogRequestBuilder}
\description{
CatalogRequestBuilder

CatalogRequestBuilder
}
\section{Super class}{
\code{\link[ArctosR:RequestBuilder]{ArctosR::RequestBuilder}} -> \code{CatalogRequestBuilder}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CatalogRequestBuilder-set_limit}{\code{CatalogRequestBuilder$set_limit()}}
\item \href{#method-CatalogRequestBuilder-set_query}{\code{CatalogRequestBuilder$set_query()}}
\item \href{#method-CatalogRequestBuilder-set_filter}{\code{CatalogRequestBuilder$set_filter()}}
\item \href{#method-CatalogRequestBuilder-set_parts}{\code{CatalogRequestBuilder$set_parts()}}
\item \href{#method-CatalogRequestBuilder-set_attributes}{\code{CatalogRequestBuilder$set_attributes()}}
\item \href{#method-CatalogRequestBuilder-set_components}{\code{CatalogRequestBuilder$set_components()}}
\item \href{#method-CatalogRequestBuilder-set_columns}{\code{CatalogRequestBuilder$set_columns()}}
\item \href{#method-CatalogRequestBuilder-set_columns_list}{\code{CatalogRequestBuilder$set_columns_list()}}
\item \href{#method-CatalogRequestBuilder-from_previous_response}{\code{CatalogRequestBuilder$from_previous_response()}}
\item \href{#method-CatalogRequestBuilder-build_request}{\code{CatalogRequestBuilder$build_request()}}
\item \href{#method-CatalogRequestBuilder-clone}{\code{CatalogRequestBuilder$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ArctosR" data-topic="RequestBuilder" data-id="debug"><a href='../../ArctosR/html/RequestBuilder.html#method-RequestBuilder-debug'><code>ArctosR::RequestBuilder$debug()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatalogRequestBuilder-set_limit"></a>}}
\if{latex}{\out{\hypertarget{method-CatalogRequestBuilder-set_limit}{}}}
\subsection{Method \code{set_limit()}}{
Sets the limit on how many records to initially request
from Arctos.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogRequestBuilder$set_limit(limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{(\code{integer(1)}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{CatalogRequestBuilder}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatalogRequestBuilder-set_query"></a>}}
\if{latex}{\out{\hypertarget{method-CatalogRequestBuilder-set_query}{}}}
\subsection{Method \code{set_query()}}{
Sets the query parameters to use to search Arctos.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogRequestBuilder$set_query(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{(\code{list}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{CatalogRequestBuilder}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatalogRequestBuilder-set_filter"></a>}}
\if{latex}{\out{\hypertarget{method-CatalogRequestBuilder-set_filter}{}}}
\subsection{Method \code{set_filter()}}{
Sets the result parameters to use to filter out results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogRequestBuilder$set_filter(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{(\code{list}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{CatalogRequestBuilder}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatalogRequestBuilder-set_parts"></a>}}
\if{latex}{\out{\hypertarget{method-CatalogRequestBuilder-set_parts}{}}}
\subsection{Method \code{set_parts()}}{
Set parts to query over.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogRequestBuilder$set_parts(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parts}}{(\code{list}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{CatalogRequestBuilder}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatalogRequestBuilder-set_attributes"></a>}}
\if{latex}{\out{\hypertarget{method-CatalogRequestBuilder-set_attributes}{}}}
\subsection{Method \code{set_attributes()}}{
Set attributes to query over.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogRequestBuilder$set_attributes(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attributes}}{(\code{list}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{CatalogRequestBuilder}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatalogRequestBuilder-set_components"></a>}}
\if{latex}{\out{\hypertarget{method-CatalogRequestBuilder-set_components}{}}}
\subsection{Method \code{set_components()}}{
Set components to query over.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogRequestBuilder$set_components(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{components}}{(\code{list}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{CatalogRequestBuilder}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatalogRequestBuilder-set_columns"></a>}}
\if{latex}{\out{\hypertarget{method-CatalogRequestBuilder-set_columns}{}}}
\subsection{Method \code{set_columns()}}{
Sets the columns in the dataframe returned by Arctos.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogRequestBuilder$set_columns(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols}}{(\code{list}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{CatalogRequestBuilder}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatalogRequestBuilder-set_columns_list"></a>}}
\if{latex}{\out{\hypertarget{method-CatalogRequestBuilder-set_columns_list}{}}}
\subsection{Method \code{set_columns_list()}}{
Sets the columns in the dataframe returned by Arctos.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogRequestBuilder$set_columns_list(l)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols}}{(\code{list}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{CatalogRequestBuilder}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatalogRequestBuilder-from_previous_response"></a>}}
\if{latex}{\out{\hypertarget{method-CatalogRequestBuilder-from_previous_response}{}}}
\subsection{Method \code{from_previous_response()}}{
Sets the columns in the dataframe returned by Arctos.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogRequestBuilder$from_previous_response(response)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response}}{a response object from a previous request}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{FromResponseRequestBuilder}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatalogRequestBuilder-build_request"></a>}}
\if{latex}{\out{\hypertarget{method-CatalogRequestBuilder-build_request}{}}}
\subsection{Method \code{build_request()}}{
Send a request for data to Arctos with parameters specified
by the other methods called on this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogRequestBuilder$build_request()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\link{Response}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatalogRequestBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CatalogRequestBuilder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogRequestBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
