% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{expand_column}
\alias{expand_column}
\title{expand_column}
\usage{
expand_column(query, column_name)
}
\arguments{
\item{query}{The query object whose records contain a column to be expanded.}

\item{column_name}{The name of the column to be expanded.}
}
\value{
Nothing.
}
\description{
Expand a column in the records if the data in that column is a JSON formatted
table.
}
