% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{get_last_response_url}
\alias{get_last_response_url}
\title{Get the last URL used by a request in a query object}
\usage{
get_last_response_url(query)
}
\arguments{
\item{query}{A query object to return the URL for}
}
\value{
The URL of the last performed request in a query object
}
\description{
Returns the last URL used by a request in a query object
}
\examples{
library(ArctosR)

query <- get_records(
  scientific_name = "Canis lupus", guid_prefix = "MSB:Mamm",
  columns = list("guid", "parts", "partdetail")
)

get_last_response_url(query)

}
