% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{get_relationships}
\alias{get_relationships}
\title{Get the relationships (e.g. "host of") a cataloged item has}
\usage{
get_relationships(guid)
}
\arguments{
\item{guid}{The Arctos GUID of the item to query relationships over}
}
\value{
a data frame of all related items. This contains URLs
}
\description{
A cataloged item in Arctos can be related to any other number of
items by relationships defined in the code table \code{ctid_references}. This
function will return all items related by any such relationship in the table
in a data frame.
}
\examples{
library(ArctosR)

if (interactive()) {
  r <- get_relationships("MSB:Mamm:140026")
}

}
