% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{get_request_url}
\alias{get_request_url}
\title{Get the first URL in a completed query}
\usage{
get_request_url(query)
}
\arguments{
\item{query}{A completed query returned from \code{get_records}}
}
\value{
A URL as a string
}
\description{
Returns the first URL used by a completed query which can be
shared. The API key is automatically stripped from the URL for security.
}
\examples{
library(ArctosR)

if (interactive()) {
  q <- get_records(guid_prefix="MSB:Mamm")
  url <- get_request_url(q)
}

}
