% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_request.R
\name{InfoRequestBuilder}
\alias{InfoRequestBuilder}
\title{InfoRequestBuilder}
\description{
Builder for a request for query parameter or result parameter documentation
from Arctos. For a valid request, only one method to specify the type of
request can be called.
}
\examples{
results <- ArctosR::InfoRequestBuilder$new()$
  all_query_params()$
  perform_request()

results <- ArctosR::InfoRequestBuilder$new()$
  all_result_params()$
  perform_request()

results <- ArctosR::InfoRequestBuilder$new()$
  by_query_category("parts")$
  perform_request()

results <- ArctosR::InfoRequestBuilder$new()$
  by_result_category("core")$
  perform_request()

results <- ArctosR::InfoRequestBuilder$new()$
  for_query_param("guid")$
  perform_request()

results <- ArctosR::InfoRequestBuilder$new()$
  for_result_param("guid")$
  perform_request()

}
\section{Super class}{
\code{\link[ArctosR:RequestBuilder]{ArctosR::RequestBuilder}} -> \code{InfoRequestBuilder}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InfoRequestBuilder-all_query_params}{\code{InfoRequestBuilder$all_query_params()}}
\item \href{#method-InfoRequestBuilder-all_result_params}{\code{InfoRequestBuilder$all_result_params()}}
\item \href{#method-InfoRequestBuilder-by_query_category}{\code{InfoRequestBuilder$by_query_category()}}
\item \href{#method-InfoRequestBuilder-by_result_category}{\code{InfoRequestBuilder$by_result_category()}}
\item \href{#method-InfoRequestBuilder-for_query_param}{\code{InfoRequestBuilder$for_query_param()}}
\item \href{#method-InfoRequestBuilder-for_result_param}{\code{InfoRequestBuilder$for_result_param()}}
\item \href{#method-InfoRequestBuilder-perform_request}{\code{InfoRequestBuilder$perform_request()}}
\item \href{#method-InfoRequestBuilder-clone}{\code{InfoRequestBuilder$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ArctosR" data-topic="RequestBuilder" data-id="debug"><a href='../../ArctosR/html/RequestBuilder.html#method-RequestBuilder-debug'><code>ArctosR::RequestBuilder$debug()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ArctosR" data-topic="RequestBuilder" data-id="default_api_key"><a href='../../ArctosR/html/RequestBuilder.html#method-RequestBuilder-default_api_key'><code>ArctosR::RequestBuilder$default_api_key()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ArctosR" data-topic="RequestBuilder" data-id="set_api_key"><a href='../../ArctosR/html/RequestBuilder.html#method-RequestBuilder-set_api_key'><code>ArctosR::RequestBuilder$set_api_key()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoRequestBuilder-all_query_params"></a>}}
\if{latex}{\out{\hypertarget{method-InfoRequestBuilder-all_query_params}{}}}
\subsection{Method \code{all_query_params()}}{
Requests all parameters that can be used as part of a query.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoRequestBuilder$all_query_params()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InfoRequestBuilder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoRequestBuilder-all_result_params"></a>}}
\if{latex}{\out{\hypertarget{method-InfoRequestBuilder-all_result_params}{}}}
\subsection{Method \code{all_result_params()}}{
Requests all columns that can be returned with a request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoRequestBuilder$all_result_params()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoRequestBuilder-by_query_category"></a>}}
\if{latex}{\out{\hypertarget{method-InfoRequestBuilder-by_query_category}{}}}
\subsection{Method \code{by_query_category()}}{
Requests all query parameters by category.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoRequestBuilder$by_query_category(category)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{Name of query category.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoRequestBuilder-by_result_category"></a>}}
\if{latex}{\out{\hypertarget{method-InfoRequestBuilder-by_result_category}{}}}
\subsection{Method \code{by_result_category()}}{
Requests all result parameters by category.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoRequestBuilder$by_result_category(category)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{Name of result category.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoRequestBuilder-for_query_param"></a>}}
\if{latex}{\out{\hypertarget{method-InfoRequestBuilder-for_query_param}{}}}
\subsection{Method \code{for_query_param()}}{
Requests a specific query parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoRequestBuilder$for_query_param(param)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{Name of query parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoRequestBuilder-for_result_param"></a>}}
\if{latex}{\out{\hypertarget{method-InfoRequestBuilder-for_result_param}{}}}
\subsection{Method \code{for_result_param()}}{
Requests a specific result parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoRequestBuilder$for_result_param(param)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{Name of result parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoRequestBuilder-perform_request"></a>}}
\if{latex}{\out{\hypertarget{method-InfoRequestBuilder-perform_request}{}}}
\subsection{Method \code{perform_request()}}{
Performs the request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoRequestBuilder$perform_request()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\link{data.frame}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoRequestBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InfoRequestBuilder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoRequestBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
