% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{Query}
\alias{Query}
\title{Query}
\description{
The results of a user query. Able to accept multiple responses
to increase the record count, or to add columns.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Query-catalog_request}{\code{Query$catalog_request()}}
\item \href{#method-Query-from_response_request}{\code{Query$from_response_request()}}
\item \href{#method-Query-info_request}{\code{Query$info_request()}}
\item \href{#method-Query-perform}{\code{Query$perform()}}
\item \href{#method-Query-save_metadata_json}{\code{Query$save_metadata_json()}}
\item \href{#method-Query-save_records_csv}{\code{Query$save_records_csv()}}
\item \href{#method-Query-expand_col}{\code{Query$expand_col()}}
\item \href{#method-Query-get_responses}{\code{Query$get_responses()}}
\item \href{#method-Query-clone}{\code{Query$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-catalog_request"></a>}}
\if{latex}{\out{\hypertarget{method-Query-catalog_request}{}}}
\subsection{Method \code{catalog_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$catalog_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-from_response_request"></a>}}
\if{latex}{\out{\hypertarget{method-Query-from_response_request}{}}}
\subsection{Method \code{from_response_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$from_response_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-info_request"></a>}}
\if{latex}{\out{\hypertarget{method-Query-info_request}{}}}
\subsection{Method \code{info_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$info_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-perform"></a>}}
\if{latex}{\out{\hypertarget{method-Query-perform}{}}}
\subsection{Method \code{perform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$perform(api_key = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-save_metadata_json"></a>}}
\if{latex}{\out{\hypertarget{method-Query-save_metadata_json}{}}}
\subsection{Method \code{save_metadata_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$save_metadata_json(file_path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-save_records_csv"></a>}}
\if{latex}{\out{\hypertarget{method-Query-save_records_csv}{}}}
\subsection{Method \code{save_records_csv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$save_records_csv(file_path, expanded = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-expand_col"></a>}}
\if{latex}{\out{\hypertarget{method-Query-expand_col}{}}}
\subsection{Method \code{expand_col()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$expand_col(column_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-get_responses"></a>}}
\if{latex}{\out{\hypertarget{method-Query-get_responses}{}}}
\subsection{Method \code{get_responses()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$get_responses()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Query-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
