% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{Records}
\alias{Records}
\title{Records}
\description{
A (possibly nested) data frame of records returned by a static
set of query and result parameters
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Records-new}{\code{Records$new()}}
\item \href{#method-Records-append}{\code{Records$append()}}
\item \href{#method-Records-save_flat_csv}{\code{Records$save_flat_csv()}}
\item \href{#method-Records-save_nested_csvs}{\code{Records$save_nested_csvs()}}
\item \href{#method-Records-expand_col}{\code{Records$expand_col()}}
\item \href{#method-Records-clone}{\code{Records$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-new"></a>}}
\if{latex}{\out{\hypertarget{method-Records-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$new(df, tbl)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-append"></a>}}
\if{latex}{\out{\hypertarget{method-Records-append}{}}}
\subsection{Method \code{append()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$append(other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-save_flat_csv"></a>}}
\if{latex}{\out{\hypertarget{method-Records-save_flat_csv}{}}}
\subsection{Method \code{save_flat_csv()}}{
Writes the data in the response object to a CSV file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$save_flat_csv(file_path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-save_nested_csvs"></a>}}
\if{latex}{\out{\hypertarget{method-Records-save_nested_csvs}{}}}
\subsection{Method \code{save_nested_csvs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$save_nested_csvs(file_path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-expand_col"></a>}}
\if{latex}{\out{\hypertarget{method-Records-expand_col}{}}}
\subsection{Method \code{expand_col()}}{
Expand a column of nested JSON tables in the response to a
list of dataframes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$expand_col(column)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{(\code{string})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Records-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
