% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{Request}
\alias{Request}
\title{Request}
\description{
A generic Arctos request. Not intended to be used directly. See
InfoRequestBuilder and CatalogRequestBuilder.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Request-with_endpoint}{\code{Request$with_endpoint()}}
\item \href{#method-Request-add_param}{\code{Request$add_param()}}
\item \href{#method-Request-add_params}{\code{Request$add_params()}}
\item \href{#method-Request-serialize}{\code{Request$serialize()}}
\item \href{#method-Request-perform}{\code{Request$perform()}}
\item \href{#method-Request-from_raw_response}{\code{Request$from_raw_response()}}
\item \href{#method-Request-clone}{\code{Request$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-with_endpoint"></a>}}
\if{latex}{\out{\hypertarget{method-Request-with_endpoint}{}}}
\subsection{Method \code{with_endpoint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$with_endpoint(endpoint)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-add_param"></a>}}
\if{latex}{\out{\hypertarget{method-Request-add_param}{}}}
\subsection{Method \code{add_param()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$add_param(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-add_params"></a>}}
\if{latex}{\out{\hypertarget{method-Request-add_params}{}}}
\subsection{Method \code{add_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$add_params(l)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-Request-serialize}{}}}
\subsection{Method \code{serialize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$serialize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-perform"></a>}}
\if{latex}{\out{\hypertarget{method-Request-perform}{}}}
\subsection{Method \code{perform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$perform(api_key = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-from_raw_response"></a>}}
\if{latex}{\out{\hypertarget{method-Request-from_raw_response}{}}}
\subsection{Method \code{from_raw_response()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$from_raw_response(raw_response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Request-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
