% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_builder.R
\name{RequestBuilder}
\alias{RequestBuilder}
\title{RequestBuilder}
\description{
A builder for a generic Arctos request. Not to be used directly.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RequestBuilder-set_api_key}{\code{RequestBuilder$set_api_key()}}
\item \href{#method-RequestBuilder-debug}{\code{RequestBuilder$debug()}}
\item \href{#method-RequestBuilder-default_api_key}{\code{RequestBuilder$default_api_key()}}
\item \href{#method-RequestBuilder-perform_request}{\code{RequestBuilder$perform_request()}}
\item \href{#method-RequestBuilder-clone}{\code{RequestBuilder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestBuilder-set_api_key"></a>}}
\if{latex}{\out{\hypertarget{method-RequestBuilder-set_api_key}{}}}
\subsection{Method \code{set_api_key()}}{
Set the API key for this request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestBuilder$set_api_key(api_key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_key}}{API key in string format.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestBuilder-debug"></a>}}
\if{latex}{\out{\hypertarget{method-RequestBuilder-debug}{}}}
\subsection{Method \code{debug()}}{
Turn on printing of debug information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestBuilder$debug()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestBuilder-default_api_key"></a>}}
\if{latex}{\out{\hypertarget{method-RequestBuilder-default_api_key}{}}}
\subsection{Method \code{default_api_key()}}{
Use the package's default API key for this request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestBuilder$default_api_key()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestBuilder-perform_request"></a>}}
\if{latex}{\out{\hypertarget{method-RequestBuilder-perform_request}{}}}
\subsection{Method \code{perform_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestBuilder$perform_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequestBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RequestBuilder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequestBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
