% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{Response}
\alias{Response}
\title{Response}
\description{
Response returned from Arctos.
}
\examples{

## ------------------------------------------------
## Method `Response$save_object`
## ------------------------------------------------

r <- ArctosR::Response$from_file("response.RData")

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Response-new}{\code{Response$new()}}
\item \href{#method-Response-was_success}{\code{Response$was_success()}}
\item \href{#method-Response-get_table_id}{\code{Response$get_table_id()}}
\item \href{#method-Response-get_api_key}{\code{Response$get_api_key()}}
\item \href{#method-Response-get_total_record_count}{\code{Response$get_total_record_count()}}
\item \href{#method-Response-get_index}{\code{Response$get_index()}}
\item \href{#method-Response-set_index}{\code{Response$set_index()}}
\item \href{#method-Response-get_record_count}{\code{Response$get_record_count()}}
\item \href{#method-Response-get_error}{\code{Response$get_error()}}
\item \href{#method-Response-get_url}{\code{Response$get_url()}}
\item \href{#method-Response-save_csv}{\code{Response$save_csv()}}
\item \href{#method-Response-save_dataframe}{\code{Response$save_dataframe()}}
\item \href{#method-Response-save_object}{\code{Response$save_object()}}
\item \href{#method-Response-get_relations}{\code{Response$get_relations()}}
\item \href{#method-Response-as_data_frame}{\code{Response$as_data_frame()}}
\item \href{#method-Response-expand_col}{\code{Response$expand_col()}}
\item \href{#method-Response-clone}{\code{Response$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-new"></a>}}
\if{latex}{\out{\hypertarget{method-Response-new}{}}}
\subsection{Method \code{new()}}{
Creates an Arctos response object from a curl response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$new(raw_response, url_params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw_response}}{(\code{list})}

\item{\code{url_params}}{(\code{list})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Response}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-was_success"></a>}}
\if{latex}{\out{\hypertarget{method-Response-was_success}{}}}
\subsection{Method \code{was_success()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$was_success()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_table_id"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_table_id}{}}}
\subsection{Method \code{get_table_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_table_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_api_key"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_api_key}{}}}
\subsection{Method \code{get_api_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_api_key()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_total_record_count"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_total_record_count}{}}}
\subsection{Method \code{get_total_record_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_total_record_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_index"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_index}{}}}
\subsection{Method \code{get_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_index()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-set_index"></a>}}
\if{latex}{\out{\hypertarget{method-Response-set_index}{}}}
\subsection{Method \code{set_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$set_index(index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_record_count"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_record_count}{}}}
\subsection{Method \code{get_record_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_record_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_error"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_error}{}}}
\subsection{Method \code{get_error()}}{
Returns any API error returned from Arctos.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_error()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_url"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_url}{}}}
\subsection{Method \code{get_url()}}{
Returns the request URL for debugging purposes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-save_csv"></a>}}
\if{latex}{\out{\hypertarget{method-Response-save_csv}{}}}
\subsection{Method \code{save_csv()}}{
Writes the data in the response object to a CSV file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$save_csv(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-save_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-Response-save_dataframe}{}}}
\subsection{Method \code{save_dataframe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$save_dataframe(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-save_object"></a>}}
\if{latex}{\out{\hypertarget{method-Response-save_object}{}}}
\subsection{Method \code{save_object()}}{
Store the entire response object as an .RData file which can
be later loaded using Response$from_file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$save_object(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(\code{string})}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{r <- ArctosR::Response$from_file("response.RData")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_relations"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_relations}{}}}
\subsection{Method \code{get_relations()}}{
Using the records in this response, request a new table
consisting of all records that are referenced in relatedcatalogeditems.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_relations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-as_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-Response-as_data_frame}{}}}
\subsection{Method \code{as_data_frame()}}{
Returns data from the response as a dataframe object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$as_data_frame()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{data.frame}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-expand_col"></a>}}
\if{latex}{\out{\hypertarget{method-Response-expand_col}{}}}
\subsection{Method \code{expand_col()}}{
Expand a column of nested JSON tables to a list of
dataframes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$expand_col(column)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{(\code{string})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Response-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
