% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{response_data}
\alias{response_data}
\title{response_data}
\usage{
response_data(query)
}
\arguments{
\item{query}{The query object to extract the dataframe from.}
}
\value{
A data frame
}
\description{
Obtain just the dataframe from a successful query.
}
