% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arctosr_doc.R
\docType{package}
\name{ArctosR-package}
\alias{ArctosR}
\alias{ArctosR-package}
\title{ArctosR: An Interface to the Arctos Database}
\description{
The \code{ArctosR} package provides a set of functions to help users
perform requests to the Arctos API to download data. It provides a set of
builder classes for performing complex requests, as well as a set of simple
functions for automating many common requests and workflows.
}
\section{About Arctos}{

Arctos is community and an online collection management information system.
Arctos is a consortium of museums and organizations that collaborate to serve
data on over 5 million records from natural and cultural history collections.
Arctos integrates access to information from diverse disciplines:
anthropology, botany, entomology, ethnology, herpetology, geology,
ichthyology, mammalogy, mineralogy, ornithology, paleontology,
parasitology as well as archival and cultural collections.
The web interface to the Arctos database can be found at
https://arctos.database.museum/. For more information about Arctos,
see https://arctosdb.org/about/, and
Cicero et al. (2024) \url{doi:10.1371/journal.pone.0296478}.
}

\section{Functions in ArctosR}{

\code{\link{get_query_parameters}}, \code{\link{get_result_parameters}},
\code{\link{get_record_count}}, \code{\link{get_records}},
\code{\link{response_data}}, \code{\link{save_response_rds}},
\code{\link{read_response_rds}}, \code{\link{save_response_csv}},
\code{\link{expand_column}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrhwilliams/arctosr}
  \item Report bugs at \url{https://github.com/hrhwilliams/arctosr/issues}
}

}
\author{
\strong{Maintainer}: Harlan Williams \email{harlanrhwilliams@gmail.com}

Authors:
\itemize{
  \item Marlon E. Cobos \email{manubio13@gmail.com}
  \item Jocelyn P. Colella \email{colella@ku.edu}
  \item Michelle S. Koo \email{mkoo@berkeley.edu}
  \item Vijay Barve \email{vijay.barve@gmail.com}
}

}
