% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_response_builder.R
\name{FromResponseRequestBuilder}
\alias{FromResponseRequestBuilder}
\title{FromResponseRequestBuilder}
\description{
Builder for the case where a request is made with the context
of a previous response.
}
\section{Super class}{
\code{\link[ArctosR:RequestBuilder]{ArctosR::RequestBuilder}} -> \code{FromResponseRequestBuilder}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FromResponseRequestBuilder-new}{\code{FromResponseRequestBuilder$new()}}
\item \href{#method-FromResponseRequestBuilder-request_more}{\code{FromResponseRequestBuilder$request_more()}}
\item \href{#method-FromResponseRequestBuilder-build_request}{\code{FromResponseRequestBuilder$build_request()}}
\item \href{#method-FromResponseRequestBuilder-clone}{\code{FromResponseRequestBuilder$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ArctosR" data-topic="RequestBuilder" data-id="debug"><a href='../../ArctosR/html/RequestBuilder.html#method-RequestBuilder-debug'><code>ArctosR::RequestBuilder$debug()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FromResponseRequestBuilder-new"></a>}}
\if{latex}{\out{\hypertarget{method-FromResponseRequestBuilder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FromResponseRequestBuilder$new(response, records)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FromResponseRequestBuilder-request_more"></a>}}
\if{latex}{\out{\hypertarget{method-FromResponseRequestBuilder-request_more}{}}}
\subsection{Method \code{request_more()}}{
Request at most \code{count} more records from this response's
original query
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FromResponseRequestBuilder$request_more(count)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{number of additional records to request}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
FromResponseRequestBuilder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FromResponseRequestBuilder-build_request"></a>}}
\if{latex}{\out{\hypertarget{method-FromResponseRequestBuilder-build_request}{}}}
\subsection{Method \code{build_request()}}{
Perform the request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FromResponseRequestBuilder$build_request()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Request
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FromResponseRequestBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FromResponseRequestBuilder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FromResponseRequestBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
