% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_response_builder.R
\name{FromResponseRequestBuilder}
\alias{FromResponseRequestBuilder}
\title{FromResponseRequestBuilder}
\description{
Builder for the case where a request is made with the context
of a previous response.
}
\examples{
response <- ArctosR::CatalogRequestBuilder$new()$
  default_api_key()$
  set_query(scientific_name="Canis lupus", guid_prefix="MSB:Mamm")$
  set_columns("guid", "parts", "partdetail")$
  set_limit(500)$
  perform_request()

response <- ArctosR::CatalogRequestBuilder$new()$
  from_previous_response(response)$
  request_more(500)$
  perform_request()

}
\section{Super class}{
\code{\link[ArctosR:RequestBuilder]{ArctosR::RequestBuilder}} -> \code{FromResponseRequestBuilder}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FromResponseRequestBuilder-new}{\code{FromResponseRequestBuilder$new()}}
\item \href{#method-FromResponseRequestBuilder-request_more}{\code{FromResponseRequestBuilder$request_more()}}
\item \href{#method-FromResponseRequestBuilder-add_column}{\code{FromResponseRequestBuilder$add_column()}}
\item \href{#method-FromResponseRequestBuilder-perform_request}{\code{FromResponseRequestBuilder$perform_request()}}
\item \href{#method-FromResponseRequestBuilder-clone}{\code{FromResponseRequestBuilder$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ArctosR" data-topic="RequestBuilder" data-id="debug"><a href='../../ArctosR/html/RequestBuilder.html#method-RequestBuilder-debug'><code>ArctosR::RequestBuilder$debug()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ArctosR" data-topic="RequestBuilder" data-id="default_api_key"><a href='../../ArctosR/html/RequestBuilder.html#method-RequestBuilder-default_api_key'><code>ArctosR::RequestBuilder$default_api_key()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ArctosR" data-topic="RequestBuilder" data-id="set_api_key"><a href='../../ArctosR/html/RequestBuilder.html#method-RequestBuilder-set_api_key'><code>ArctosR::RequestBuilder$set_api_key()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FromResponseRequestBuilder-new"></a>}}
\if{latex}{\out{\hypertarget{method-FromResponseRequestBuilder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FromResponseRequestBuilder$new(response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FromResponseRequestBuilder-request_more"></a>}}
\if{latex}{\out{\hypertarget{method-FromResponseRequestBuilder-request_more}{}}}
\subsection{Method \code{request_more()}}{
Request at most \code{count} more records from this response's
original query
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FromResponseRequestBuilder$request_more(count)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{number of additional records to request}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
FromResponseRequestBuilder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FromResponseRequestBuilder-add_column"></a>}}
\if{latex}{\out{\hypertarget{method-FromResponseRequestBuilder-add_column}{}}}
\subsection{Method \code{add_column()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FromResponseRequestBuilder$add_column(col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FromResponseRequestBuilder-perform_request"></a>}}
\if{latex}{\out{\hypertarget{method-FromResponseRequestBuilder-perform_request}{}}}
\subsection{Method \code{perform_request()}}{
Perform the request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FromResponseRequestBuilder$perform_request()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Response
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FromResponseRequestBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FromResponseRequestBuilder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FromResponseRequestBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
