% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_request.R
\name{InfoRequestBuilder}
\alias{InfoRequestBuilder}
\title{InfoRequestBuilder}
\description{
Builder for a request for query parameter or result parameter documentation
from Arctos. For a valid request, only one method to specify the type of
request can be called.
}
\section{Super class}{
\code{\link[ArctosR:RequestBuilder]{ArctosR::RequestBuilder}} -> \code{InfoRequestBuilder}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InfoRequestBuilder-build_request}{\code{InfoRequestBuilder$build_request()}}
\item \href{#method-InfoRequestBuilder-clone}{\code{InfoRequestBuilder$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ArctosR" data-topic="RequestBuilder" data-id="debug"><a href='../../ArctosR/html/RequestBuilder.html#method-RequestBuilder-debug'><code>ArctosR::RequestBuilder$debug()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoRequestBuilder-build_request"></a>}}
\if{latex}{\out{\hypertarget{method-InfoRequestBuilder-build_request}{}}}
\subsection{Method \code{build_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoRequestBuilder$build_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfoRequestBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InfoRequestBuilder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfoRequestBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
