% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{Metadata}
\alias{Metadata}
\title{Metadata}
\description{
Metadata for a specific HTTP response from Arctos.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Metadata-to_list}{\code{Metadata$to_list()}}
\item \href{#method-Metadata-clone}{\code{Metadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
