% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{Response}
\alias{Response}
\title{Response}
\description{
Response returned from Arctos.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Response-new}{\code{Response$new()}}
\item \href{#method-Response-set_start_index}{\code{Response$set_start_index()}}
\item \href{#method-Response-was_success}{\code{Response$was_success()}}
\item \href{#method-Response-is_empty}{\code{Response$is_empty()}}
\item \href{#method-Response-has_json_content}{\code{Response$has_json_content()}}
\item \href{#method-Response-to_list}{\code{Response$to_list()}}
\item \href{#method-Response-to_records}{\code{Response$to_records()}}
\item \href{#method-Response-clone}{\code{Response$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-new"></a>}}
\if{latex}{\out{\hypertarget{method-Response-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$new(request, raw_response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-set_start_index"></a>}}
\if{latex}{\out{\hypertarget{method-Response-set_start_index}{}}}
\subsection{Method \code{set_start_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$set_start_index(start)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-was_success"></a>}}
\if{latex}{\out{\hypertarget{method-Response-was_success}{}}}
\subsection{Method \code{was_success()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$was_success()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-is_empty"></a>}}
\if{latex}{\out{\hypertarget{method-Response-is_empty}{}}}
\subsection{Method \code{is_empty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$is_empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-has_json_content"></a>}}
\if{latex}{\out{\hypertarget{method-Response-has_json_content}{}}}
\subsection{Method \code{has_json_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$has_json_content()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Response-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-to_records"></a>}}
\if{latex}{\out{\hypertarget{method-Response-to_records}{}}}
\subsection{Method \code{to_records()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$to_records(start = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Response-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
