% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{Response}
\alias{Response}
\title{Response}
\description{
Response returned from Arctos.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Response-new}{\code{Response$new()}}
\item \href{#method-Response-get_metadata}{\code{Response$get_metadata()}}
\item \href{#method-Response-was_success}{\code{Response$was_success()}}
\item \href{#method-Response-get_table_id}{\code{Response$get_table_id()}}
\item \href{#method-Response-get_api_key}{\code{Response$get_api_key()}}
\item \href{#method-Response-get_total_record_count}{\code{Response$get_total_record_count()}}
\item \href{#method-Response-get_index}{\code{Response$get_index()}}
\item \href{#method-Response-set_index}{\code{Response$set_index()}}
\item \href{#method-Response-get_record_count}{\code{Response$get_record_count()}}
\item \href{#method-Response-get_error}{\code{Response$get_error()}}
\item \href{#method-Response-get_url}{\code{Response$get_url()}}
\item \href{#method-Response-save_flat_csv}{\code{Response$save_flat_csv()}}
\item \href{#method-Response-save_expanded_csvs}{\code{Response$save_expanded_csvs()}}
\item \href{#method-Response-save_object}{\code{Response$save_object()}}
\item \href{#method-Response-as_data_frame}{\code{Response$as_data_frame()}}
\item \href{#method-Response-expand_col}{\code{Response$expand_col()}}
\item \href{#method-Response-clone}{\code{Response$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-new"></a>}}
\if{latex}{\out{\hypertarget{method-Response-new}{}}}
\subsection{Method \code{new()}}{
Creates an Arctos response object from a curl response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$new(raw_response, url_params, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw_response}}{(\code{list})}

\item{\code{url_params}}{(\code{list})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Response}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-was_success"></a>}}
\if{latex}{\out{\hypertarget{method-Response-was_success}{}}}
\subsection{Method \code{was_success()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$was_success()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_table_id"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_table_id}{}}}
\subsection{Method \code{get_table_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_table_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_api_key"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_api_key}{}}}
\subsection{Method \code{get_api_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_api_key()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_total_record_count"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_total_record_count}{}}}
\subsection{Method \code{get_total_record_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_total_record_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_index"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_index}{}}}
\subsection{Method \code{get_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_index()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-set_index"></a>}}
\if{latex}{\out{\hypertarget{method-Response-set_index}{}}}
\subsection{Method \code{set_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$set_index(index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_record_count"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_record_count}{}}}
\subsection{Method \code{get_record_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_record_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_error"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_error}{}}}
\subsection{Method \code{get_error()}}{
Returns any API error returned from Arctos.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_error()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_url"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_url}{}}}
\subsection{Method \code{get_url()}}{
Returns the request URL for debugging purposes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-save_flat_csv"></a>}}
\if{latex}{\out{\hypertarget{method-Response-save_flat_csv}{}}}
\subsection{Method \code{save_flat_csv()}}{
Writes the data in the response object to a CSV file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$save_flat_csv(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-save_expanded_csvs"></a>}}
\if{latex}{\out{\hypertarget{method-Response-save_expanded_csvs}{}}}
\subsection{Method \code{save_expanded_csvs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$save_expanded_csvs(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-save_object"></a>}}
\if{latex}{\out{\hypertarget{method-Response-save_object}{}}}
\subsection{Method \code{save_object()}}{
Store the entire response object as an .RData file which can
be later loaded using Response$from_file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$save_object(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(\code{string})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-as_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-Response-as_data_frame}{}}}
\subsection{Method \code{as_data_frame()}}{
Returns data from the response as a dataframe object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$as_data_frame()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{data.frame}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-expand_col"></a>}}
\if{latex}{\out{\hypertarget{method-Response-expand_col}{}}}
\subsection{Method \code{expand_col()}}{
Expand a column of nested JSON tables in the response to a
list of dataframes. This does not modify the original data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$expand_col(column)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{(\code{string})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Response-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
