% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{Response}
\alias{Response}
\title{Response}
\description{
Response returned from Arctos.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Response-new}{\code{Response$new()}}
\item \href{#method-Response-request_more}{\code{Response$request_more()}}
\item \href{#method-Response-get_error}{\code{Response$get_error()}}
\item \href{#method-Response-get_url}{\code{Response$get_url()}}
\item \href{#method-Response-to_csv}{\code{Response$to_csv()}}
\item \href{#method-Response-get_relations}{\code{Response$get_relations()}}
\item \href{#method-Response-as_data_frame}{\code{Response$as_data_frame()}}
\item \href{#method-Response-expand_col}{\code{Response$expand_col()}}
\item \href{#method-Response-clone}{\code{Response$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-new"></a>}}
\if{latex}{\out{\hypertarget{method-Response-new}{}}}
\subsection{Method \code{new()}}{
Creates an Arctos response object from a curl response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$new(raw_response, url_params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw_response}}{(\code{list})}

\item{\code{url_params}}{(\code{list})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Response}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-request_more"></a>}}
\if{latex}{\out{\hypertarget{method-Response-request_more}{}}}
\subsection{Method \code{request_more()}}{
Requests more records from Arctos.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$request_more(count)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{(\code{integer})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Response}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_error"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_error}{}}}
\subsection{Method \code{get_error()}}{
Returns any API error returned from Arctos.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_error()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_url"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_url}{}}}
\subsection{Method \code{get_url()}}{
Returns the request URL for debugging purposes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-to_csv"></a>}}
\if{latex}{\out{\hypertarget{method-Response-to_csv}{}}}
\subsection{Method \code{to_csv()}}{
Writes the data in the response object to a CSV file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$to_csv(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-get_relations"></a>}}
\if{latex}{\out{\hypertarget{method-Response-get_relations}{}}}
\subsection{Method \code{get_relations()}}{
Using the records in this response, request a new table
consisting of all records that are referenced in relatedcatalogeditems
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$get_relations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-as_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-Response-as_data_frame}{}}}
\subsection{Method \code{as_data_frame()}}{
Returns data from the response as a dataframe object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$as_data_frame()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{data.frame}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-expand_col"></a>}}
\if{latex}{\out{\hypertarget{method-Response-expand_col}{}}}
\subsection{Method \code{expand_col()}}{
Expand nested JSON objects to dataframes in place
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$expand_col(column)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{(\code{string})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Response-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
